# src/main/java/it/pagopa/afm/marketplacebe/model/PageInfo.java
package it.pagopa.afm.marketplacebe.model;

import com.fasterxml.jackson.annotation.JsonAlias;
import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.v3.oas.annotations.media.Schema;
import javax.validation.constraints.Positive;
import javax.validation.constraints.PositiveOrZero;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.ToString;

@Data
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor
@ToString
@JsonIgnoreProperties(ignoreUnknown = true)
public class PageInfo {

  @JsonProperty("page")
  @Schema(description = "Page number", required = true)
  @PositiveOrZero
  Integer page;

  @JsonProperty("limit")
  @Schema(description = "Required number of items per page", required = true)
  @Positive
  Integer limit;

  @JsonProperty("itemsFound")
  @Schema(
      description = "Number of items found. (The last page may have fewer elements than required)",
      required = true)
  @PositiveOrZero
  Integer itemsFound;

  @JsonProperty("totalPages")
  @Schema(description = "Total number of pages", required = true)
  @PositiveOrZero
  Integer totalPages;

  @JsonProperty("total_items")
  @JsonAlias("totalItems")
  Long totalItems;
}