# src/main/java/it/pagopa/afm/marketplacebe/model/touchpoint/Touchpoint.java
package it.pagopa.afm.marketplacebe.model.touchpoint;

import java.time.LocalDateTime;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.experimental.SuperBuilder;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@SuperBuilder(toBuilder = true)
public class Touchpoint {

  private String id;
  private String name;
  private LocalDateTime createdDate;
}